refactor(stella-cli): split fleet_cmd's reporting into a sibling
#4801 left fleet_cmd.rs at 1499 of 1500 lines and #4813 recorded that the
next line added anywhere in it would fail the gate. That line landed and
the file reached 1538.

The reporting slice — truncate, print_dash_summary, render_report — is the
part with the fewest ties to the run loop: it reads a finished
FleetRunReport and writes to stdout. Moved verbatim into
fleet_cmd/report.rs; no test deleted or renamed.

fleet_cmd.rs is 1425 lines now, so it has real headroom rather than one
line of it.

Closes #4821
Closes #4813
